+++ CHANGELOG.md
@@ -25,6 +25,14 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.0] - 2018-02-06
 
+## [1.4.7] - 2018-02-20
+
+### Fixed
+
+- parentEvent id must not equal event id
+- back button in epid list view fixed
+- improved event filter for childEvents
+
 ### Changed
 - `heimrichhannot/contao-formhybrid` 3.x dependency